+++ nf_core/lint.py
@@ -170,7 +170,6 @@ class PipelineLint(object):
             'check_actions_branch_protection',
             'check_actions_ci',
             'check_actions_lint',
-            'check_ci_config',
             'check_readme',
             'check_conda_env_yaml',
             'check_conda_dockerfile',
@@ -237,18 +236,14 @@ class PipelineLint(object):
             os.path.join('docs','README.md'),
             os.path.join('docs','output.md'),
             os.path.join('docs','usage.md'),
-            ['.travis.yml', os.path.join('.github', 'workflows', 'branch.yml')],
-            ['.travis.yml', os.path.join('.github', 'workflows','ci.yml')],
-            ['.travis.yml', os.path.join('.github', 'workflows', 'linting.yml')]
-
+            os.path.join('.github', 'workflows', 'branch.yml'),
+            os.path.join('.github', 'workflows','ci.yml'),
+            os.path.join('.github', 'workflows', 'linting.yml')
         ]
         files_warn = [
             'main.nf',
             'environment.yml',
-            os.path.join('conf','base.config'),
-            os.path.join('.github', 'workflows', 'branch.yml'),
-            os.path.join('.github', 'workflows','ci.yml'),
-            os.path.join('.github', 'workflows', 'linting.yml')
+            os.path.join('conf','base.config')
         ]
         files_fail_ifexists = [
             'Singularity'
@@ -490,7 +485,7 @@ class PipelineLint(object):
                 self.passed.append((4, "Config variable process.container looks correct: '{}'".format(container_name)))
 
     def check_actions_branch_protection(self):
-        """Checks that the GitHub actions branch protection workflow is valid.
+        """Checks that the GitHub Actions branch protection workflow is valid.
 
         Makes sure PRs can only come from nf-core dev or 'patch' of a fork.
         """
@@ -503,9 +498,9 @@ class PipelineLint(object):
             try:
                 assert('master' in branchwf[True]['pull_request']['branches'])
             except (AssertionError, KeyError):
-                self.failed.append((5, "GitHub actions branch workflow must check for master branch PRs: '{}'".format(fn)))
+                self.failed.append((5, "GitHub Actions branch workflow must check for master branch PRs: '{}'".format(fn)))
             else:
-                self.passed.append((5, "GitHub actions branch workflow checks for master branch PRs: '{}'".format(fn)))
+                self.passed.append((5, "GitHub Actions branch workflow checks for master branch PRs: '{}'".format(fn)))
 
             # Check that PRs are only ok if coming from an nf-core `dev` branch or a fork `patch` branch
             pipeline_version = self.config.get('manifest.version', '').strip(' \'"')
@@ -515,12 +510,12 @@ class PipelineLint(object):
                 steps = branchwf['jobs']['test']['steps']
                 assert(any([PRMasterCheck in step['run'] for step in steps]))
             except (AssertionError, KeyError):
-                self.failed.append((5, "GitHub actions branch workflow checks for master branch PRs: '{}'".format(fn)))
+                self.failed.append((5, "GitHub Actions branch workflow checks for master branch PRs: '{}'".format(fn)))
             else:
-                self.passed.append((5, "GitHub actions branch workflow checks for master branch PRs: '{}'".format(fn)))
+                self.passed.append((5, "GitHub Actions branch workflow checks for master branch PRs: '{}'".format(fn)))
 
     def check_actions_ci(self):
-        """Checks that the GitHub actions ci workflow is valid
+        """Checks that the GitHub Actions CI workflow is valid
 
         Makes sure tests run with the required nextflow version.
         """
@@ -534,22 +529,31 @@ class PipelineLint(object):
                 assert('push' in ciwf[True])
                 assert('pull_request' in ciwf[True])
             except (AssertionError, KeyError, TypeError):
-                self.failed.append((5, "GitHub actions ci workflow must be triggered on PR and push: '{}'".format(fn)))
+                self.failed.append((5, "GitHub Actions CI workflow must be triggered on PR and push: '{}'".format(fn)))
             else:
-                self.passed.append((5, "GitHub actions ci workflow is triggered on PR and push: '{}'".format(fn)))
+                self.passed.append((5, "GitHub Actions CI workflow is triggered on PR and push: '{}'".format(fn)))
 
             # Check that we're pulling the right docker image and tagging it properly
             if self.config.get('process.container', ''):
                 docker_notag = re.sub(r':(?:[\.\d]+|dev)$', '', self.config.get('process.container', '').strip('"\''))
                 docker_withtag = self.config.get('process.container', '').strip('"\'')
-                docker_pull_cmd = 'docker pull {}:dev && docker tag {}:dev {}\n'.format(docker_notag, docker_notag, docker_withtag)
+                docker_pull_cmd = 'docker pull {}:dev'.format(docker_notag)
                 try:
                     steps = ciwf['jobs']['test']['steps']
                     assert(any([docker_pull_cmd in step['run'] for step in steps if 'run' in step.keys()]))
                 except (AssertionError, KeyError, TypeError):
-                    self.failed.append((5, "CI is not pulling and tagging the correct docker image. Should be:\n    '{}'".format(docker_pull_cmd)))
+                    self.failed.append((5, "CI is not pulling the correct docker image. Should be:\n    '{}'".format(docker_pull_cmd)))
                 else:
-                    self.passed.append((5, "CI is pulling and tagging the correct docker image: {}".format(docker_pull_cmd)))
+                    self.passed.append((5, "CI is pulling the correct docker image: {}".format(docker_pull_cmd)))
+
+                docker_tag_cmd = 'docker tag {}:dev {}'.format(docker_notag, docker_withtag)
+                try:
+                    steps = ciwf['jobs']['test']['steps']
+                    assert(any([docker_tag_cmd in step['run'] for step in steps if 'run' in step.keys()]))
+                except (AssertionError, KeyError, TypeError):
+                    self.failed.append((5, "CI is not tagging docker image correctly. Should be:\n    '{}'".format(docker_tag_cmd)))
+                else:
+                    self.passed.append((5, "CI is tagging docker image correctly: {}".format(docker_tag_cmd)))
 
             # Check that we are testing the minimum nextflow version
             try:
@@ -563,7 +567,7 @@ class PipelineLint(object):
                 self.passed.append((5, "Continuous integration checks minimum NF version: '{}'".format(fn)))
 
     def check_actions_lint(self):
-        """Checks that the GitHub actions lint workflow is valid
+        """Checks that the GitHub Actions lint workflow is valid
 
         Makes sure ``nf-core lint`` and ``markdownlint`` runs.
         """
@@ -577,9 +581,9 @@ class PipelineLint(object):
                 assert('push' in lintwf[True])
                 assert('pull_request' in lintwf[True])
             except (AssertionError, KeyError, TypeError):
-                self.failed.append((5, "GitHub actions linting workflow must be triggered on PR and push: '{}'".format(fn)))
+                self.failed.append((5, "GitHub Actions linting workflow must be triggered on PR and push: '{}'".format(fn)))
             else:
-                self.passed.append((5, "GitHub actions linting workflow is triggered on PR and push: '{}'".format(fn)))
+                self.passed.append((5, "GitHub Actions linting workflow is triggered on PR and push: '{}'".format(fn)))
 
             # Check that the Markdown linting runs
             Markdownlint_cmd = 'markdownlint ${GITHUB_WORKSPACE} -c ${GITHUB_WORKSPACE}/.github/markdownlint.yml'
@@ -601,73 +605,6 @@ class PipelineLint(object):
                 self.failed.append((5, "Continuous integration must run nf-core lint Tests: '{}'".format(fn)))
             else:
                 self.passed.append((5, "Continuous integration runs nf-core lint Tests: '{}'".format(fn)))
-
-    def check_ci_config(self):
-        """Checks that the Travis CI YAML config is valid.
-
-        Makes sure that ``nf-core lint`` runs in travis tests and that
-        tests run with the required nextflow version.
-        """
-        for cf in ['.travis.yml']:
-            fn = os.path.join(self.path, cf)
-            if os.path.isfile(fn):
-                with open(fn, 'r') as fh:
-                    ciconf = yaml.safe_load(fh)
-                # Check that we have the master branch protection, but allow patch as well
-                travisMasterCheck = '[ $TRAVIS_PULL_REQUEST = "false" ] || [ $TRAVIS_BRANCH != "master" ] || ([ $TRAVIS_PULL_REQUEST_SLUG = $TRAVIS_REPO_SLUG ] && [ $TRAVIS_PULL_REQUEST_BRANCH = "dev" ]) || [ $TRAVIS_PULL_REQUEST_BRANCH = "patch" ]'
-
-                try:
-                    assert(travisMasterCheck in ciconf.get('before_install', {}))
-                except AssertionError:
-                    self.failed.append((5, "Continuous integration must check for master/patch branch PRs: '{}'".format(fn)))
-                else:
-                    self.passed.append((5, "Continuous integration checks for master/patch branch PRs: '{}'".format(fn)))
-                # Check that the nf-core linting runs
-                try:
-                    assert('nf-core lint ${TRAVIS_BUILD_DIR}' in ciconf['script'])
-                except AssertionError:
-                    self.failed.append((5, "Continuous integration must run nf-core lint Tests: '{}'".format(fn)))
-                else:
-                    self.passed.append((5, "Continuous integration runs nf-core lint Tests: '{}'".format(fn)))
-
-                # Check that we're pulling the right docker image
-                if self.config.get('process.container', ''):
-                    docker_notag = re.sub(r':(?:[\.\d]+|dev)$', '', self.config.get('process.container', '').strip('"\''))
-                    docker_pull_cmd = 'docker pull {}:dev'.format(docker_notag)
-                    try:
-                        assert(docker_pull_cmd in ciconf.get('before_install', []))
-                    except AssertionError:
-                        self.failed.append((5, "CI is not pulling the correct docker image. Should be:\n    '{}'".format(docker_pull_cmd)))
-                    else:
-                        self.passed.append((5, "CI is pulling the correct docker image: {}".format(docker_pull_cmd)))
-
-                    # Check that we tag the docker image properly
-                    docker_tag_cmd = 'docker tag {}:dev {}'.format(docker_notag, self.config.get('process.container', '').strip('"\''))
-                    try:
-                        assert(docker_tag_cmd in ciconf.get('before_install'))
-                    except AssertionError:
-                        self.failed.append((5, "CI is not tagging docker image correctly. Should be:\n    '{}'".format(docker_tag_cmd)))
-                    else:
-                        self.passed.append((5, "CI is tagging docker image correctly: {}".format(docker_tag_cmd)))
-
-                # Check that we're testing the minimum nextflow version
-                minNextflowVersion = ""
-                env = ciconf.get('env', [])
-                if type(env) is dict:
-                    env = env.get('matrix', [])
-                for e in env:
-                    # Split using shlex so that we don't split "quoted whitespace"
-                    for s in shlex.split(e):
-                        k,v = s.split('=')
-                        if k == 'NXF_VER':
-                            ci_ver = v.strip('\'"')
-                            minNextflowVersion = ci_ver if v else minNextflowVersion
-                            if ci_ver == self.minNextflowVersion:
-                                self.passed.append((5, "Continuous integration checks minimum NF version: '{}'".format(fn)))
-                if not minNextflowVersion:
-                    self.failed.append((5, "Continuous integration does not check minimum NF version: '{}'".format(fn)))
-                elif minNextflowVersion != self.minNextflowVersion:
-                    self.failed.append((5, "Minimum NF version differed from CI and what was set in the pipelines manifest: {}".format(fn)))
 
     def check_readme(self):
         """Checks the repository README file for errors.
@@ -704,7 +641,7 @@ class PipelineLint(object):
     def check_version_consistency(self):
         """Checks container tags versions.
 
-        Runs on ``process.container``, ``process.container`` and ``$TRAVIS_TAG`` (each only if set).
+        Runs on ``process.container`` (if set) and ``$GITHUB_REF`` (if a GitHub Actions release).
 
         Checks that:
             * the container has a tag
@@ -729,9 +666,9 @@ class PipelineLint(object):
         if self.config.get('process.container', ''):
             versions['process.container'] = self.config.get('process.container', '').strip(' \'"').split(':')[-1]
 
-        # Get version from the TRAVIS_TAG env var
-        if os.environ.get('TRAVIS_TAG') and os.environ.get('TRAVIS_REPO_SLUG', '') != 'nf-core/tools':
-            versions['TRAVIS_TAG'] = os.environ.get('TRAVIS_TAG').strip(' \'"')
+        # Get version from the GITHUB_REF env var if this is a release
+        if os.environ.get('GITHUB_REF', '').startswith('refs/tags/') and os.environ.get('GITHUB_REPOSITORY', '') != 'nf-core/tools':
+            versions['GITHUB_REF'] = os.path.basename(os.environ['GITHUB_REF'].strip(' \'"'))
 
         # Check if they are all numeric
         for v_type, version in versions.items():